refactor : 매직넘버 상수로 변경 및 패턴 컴파일 상수로 선언
