+++ src/test/java/calculator/StringAddCalculatorTest.java
@@ -1,4 +1,4 @@
-package study;
+package calculator;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
